dvk safetyEquipmentStyles refactor
