+++ src/components/layerStyles/safetyEquipmentIcons.ts
@@ -1,0 +1,121 @@
+import a from '../../theme/img/safetyequipment/a.svg';
+import b from '../../theme/img/safetyequipment/b.svg';
+import c from '../../theme/img/safetyequipment/c.svg';
+import d from '../../theme/img/safetyequipment/d.svg';
+import e from '../../theme/img/safetyequipment/e.svg';
+import f from '../../theme/img/safetyequipment/f.svg';
+import g from '../../theme/img/safetyequipment/g.svg';
+import h from '../../theme/img/safetyequipment/h.svg';
+import i from '../../theme/img/safetyequipment/i.svg';
+import j from '../../theme/img/safetyequipment/j.svg';
+import k from '../../theme/img/safetyequipment/k.svg';
+import l from '../../theme/img/safetyequipment/l.svg';
+import m from '../../theme/img/safetyequipment/m.svg';
+import n from '../../theme/img/safetyequipment/n.svg';
+import o from '../../theme/img/safetyequipment/o.svg';
+import p from '../../theme/img/safetyequipment/p.svg';
+import q from '../../theme/img/safetyequipment/q.svg';
+import r from '../../theme/img/safetyequipment/r.svg';
+import s from '../../theme/img/safetyequipment/s.svg';
+import t from '../../theme/img/safetyequipment/t.svg';
+import u from '../../theme/img/safetyequipment/u.svg';
+import v from '../../theme/img/safetyequipment/v.svg';
+import w from '../../theme/img/safetyequipment/w.svg';
+import n1 from '../../theme/img/safetyequipment/1.svg';
+import n2 from '../../theme/img/safetyequipment/2.svg';
+import n3 from '../../theme/img/safetyequipment/3.svg';
+import n4 from '../../theme/img/safetyequipment/4.svg';
+import n5 from '../../theme/img/safetyequipment/5.svg';
+import n6 from '../../theme/img/safetyequipment/6.svg';
+import n7 from '../../theme/img/safetyequipment/7.svg';
+import n8 from '../../theme/img/safetyequipment/8.svg';
+import n9 from '../../theme/img/safetyequipment/9.svg';
+import n0 from '../../theme/img/safetyequipment/0.svg';
+import questionmark from '../../theme/img/safetyequipment/questionmark.svg';
+import A from '../../theme/img/safetyequipment/big/A.svg';
+import B from '../../theme/img/safetyequipment/big/B.svg';
+import C from '../../theme/img/safetyequipment/big/C.svg';
+import D from '../../theme/img/safetyequipment/big/D.svg';
+import E from '../../theme/img/safetyequipment/big/E.svg';
+import F from '../../theme/img/safetyequipment/big/F.svg';
+import G from '../../theme/img/safetyequipment/big/G.svg';
+import H from '../../theme/img/safetyequipment/big/H.svg';
+import I from '../../theme/img/safetyequipment/big/I.svg';
+import J from '../../theme/img/safetyequipment/big/J.svg';
+import K from '../../theme/img/safetyequipment/big/K.svg';
+import L from '../../theme/img/safetyequipment/big/L.svg';
+import M from '../../theme/img/safetyequipment/big/M.svg';
+import N from '../../theme/img/safetyequipment/big/N.svg';
+import O from '../../theme/img/safetyequipment/big/O.svg';
+import P from '../../theme/img/safetyequipment/big/P.svg';
+import Q from '../../theme/img/safetyequipment/big/Q.svg';
+import R from '../../theme/img/safetyequipment/big/R.svg';
+import S from '../../theme/img/safetyequipment/big/S.svg';
+import T from '../../theme/img/safetyequipment/big/T.svg';
+import U from '../../theme/img/safetyequipment/big/U.svg';
+import V from '../../theme/img/safetyequipment/big/V.svg';
+import W from '../../theme/img/safetyequipment/big/W.svg';
+import X from '../../theme/img/safetyequipment/big/X.svg';
+import Y from '../../theme/img/safetyequipment/big/Y.svg';
+
+export const symbol2Icon = {
+  a: { icon: a, center: false, anchorY: 32 },
+  b: { icon: b, center: false, anchorY: 32 },
+  c: { icon: c, center: false, anchorY: 32 },
+  d: { icon: d, center: false, anchorY: 32 },
+  e: { icon: e, center: false, anchorY: 32 },
+  f: { icon: f, center: false, anchorY: 32 },
+  g: { icon: g, center: false, anchorY: 32 },
+  h: { icon: h, center: false, anchorY: 32 },
+  i: { icon: i, center: false, anchorY: 32 },
+  j: { icon: j, center: false, anchorY: 32 },
+  k: { icon: k, center: false, anchorY: 32 },
+  l: { icon: l, center: false, anchorY: 32 },
+  m: { icon: m, center: false, anchorY: 32 },
+  n: { icon: n, center: false, anchorY: 32 },
+  o: { icon: o, center: false, anchorY: 32 },
+  p: { icon: p, center: false, anchorY: 32 },
+  q: { icon: q, center: false, anchorY: 32 },
+  r: { icon: r, center: false, anchorY: 32 },
+  s: { icon: s, center: false, anchorY: 32 },
+  t: { icon: t, center: true, anchorY: 0 },
+  u: { icon: u, center: false, anchorY: 29 },
+  v: { icon: v, center: false, anchorY: 32 },
+  w: { icon: w, center: false, anchorY: 32 },
+  A: { icon: A, center: false, anchorY: 29 },
+  B: { icon: B, center: false, anchorY: 29 },
+  C: { icon: C, center: false, anchorY: 29 },
+  D: { icon: D, center: false, anchorY: 29 },
+  E: { icon: E, center: false, anchorY: 29 },
+  F: { icon: F, center: false, anchorY: 29 },
+  G: { icon: G, center: false, anchorY: 32 },
+  H: { icon: H, center: false, anchorY: 32 },
+  I: { icon: I, center: false, anchorY: 32 },
+  J: { icon: J, center: false, anchorY: 32 },
+  K: { icon: K, center: false, anchorY: 32 },
+  L: { icon: L, center: false, anchorY: 32 },
+  M: { icon: M, center: false, anchorY: 32 },
+  N: { icon: N, center: false, anchorY: 32 },
+  O: { icon: O, center: false, anchorY: 32 },
+  P: { icon: P, center: false, anchorY: 32 },
+  Q: { icon: Q, center: false, anchorY: 32 },
+  R: { icon: R, center: false, anchorY: 32 },
+  S: { icon: S, center: false, anchorY: 32 },
+  T: { icon: T, center: false, anchorY: 32 },
+  U: { icon: U, center: false, anchorY: 32 },
+  V: { icon: V, center: false, anchorY: 32 },
+  W: { icon: W, center: false, anchorY: 32 },
+  X: { icon: X, center: false, anchorY: 32 },
+  Y: { icon: Y, center: false, anchorY: 32 },
+  '1': { icon: n1, center: true, anchorY: 0 },
+  '2': { icon: n2, center: false, anchorY: 29 },
+  '3': { icon: n3, center: false, anchorY: 29 },
+  '4': { icon: n4, center: false, anchorY: 29 },
+  '5': { icon: n5, center: false, anchorY: 29 },
+  '6': { icon: n6, center: false, anchorY: 29 },
+  '7': { icon: n7, center: false, anchorY: 29 },
+  '8': { icon: n8, center: false, anchorY: 29 },
+  '9': { icon: n9, center: false, anchorY: 29 },
+  '0': { icon: n0, center: false, anchorY: 29 },
+  '?': { icon: questionmark, center: false, anchorY: 28 },
+};
